+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- php: refresh config
 - ruby: refresh config
 - python: refresh config
 - java-spring: refresh config